requirements.txt
flask
flask-restful
flask-httpauth
flask-cors
python-dotenv
requests==2.26.0
boto3==1.19.0
pandas==1.3.3
xlsxwriter
Flask-BasicAuth
flask-login